@@ -3,8 +3,9 @@
   "version": "2.3.25",
   "private": true,
   "dependencies": {
-    "@openware/components": "2.0.46",
-    "@openware/cryptofont": "^0.1.6",
+    "@openware/components": "2.3.0",
+    "@openware/cryptofont": "2.3.0",
+    "@types/ws": "^6.0.4",
     "axios": "^0.18.0",
     "bitcoincashjs": "0.1.10",
     "bootstrap": "^4.3.1",
@@ -20,19 +21,20 @@
     "react-fade-in": "^0.1.6",
     "react-ga": "^2.6.0",
     "react-google-recaptcha": "^1.0.4",
-    "react-grid-layout": "^0.16.6",
+    "react-grid-layout": "^0.17.1",
     "react-intl": "^2.8.0",
     "react-maskinput": "^1.0.1",
     "react-redux": "^5.0.7",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
-    "react-scripts-ts": "^2.17.0",
+    "react-scripts": "3.0.1",
+    "recharts": "^1.8.5",
     "redux": "^4.0.0",
     "redux-saga": "^0.16.2",
     "sinon": "^7.2.3",
-    "tslint-eslint-rules": "^5.4.0",
+    "tslint": "^5.13.1",
     "ua-parser-js": "^0.7.19",
-    "webpack": "^3.11.0"
+    "ws": "^7.2.0"
   },
   "scripts": {
     "build:css": "postcss src/index.pcss -o src/index.css",
@@ -40,20 +42,20 @@
     "lint": "tsc --noEmit && tslint --project .",
     "lint:fix": "tslint --project . --fix",
     "start": "npm-run-all -p watch:css start-js",
-    "start-js": "react-app-rewired start --scripts-version react-scripts-ts start",
-    "build-js": "react-app-rewired build --scripts-version react-scripts-ts build",
+    "start-js": "react-scripts start",
+    "build-js": "react-scripts build",
     "build": "npm-run-all build:css build-js clean:mapfiles",
     "clean:mapfiles": "find build -name '*.map' -exec rm {} ';'",
-    "test": "jest",
-    "test:coverage": "jest --collectCoverage=true",
-    "test:ci": "jest --ci",
+    "test": "react-scripts test",
+    "test:coverage": "react-scripts test --collectCoverage=true",
+    "test:ci": "react-scripts test --ci",
     "spec:integration": "cypress open",
-    "eject": "react-scripts-ts eject",
+    "eject": "react-scripts eject",
     "clean": "rimraf build",
     "analyze": "source-map-explorer build/static/js/main.*"
   },
   "devDependencies": {
-    "@openware/coding-standards": "0.1.6",
+    "@openware/coding-standards": "2.3.0",
     "@types/axios": "^0.14.0",
     "@types/classnames": "^2.2.6",
     "@types/enzyme": "^3.1.13",
@@ -72,21 +74,18 @@
     "@types/webpack-env": "^1.13.6",
     "autoprefixer": "^9.1.1",
     "axios-mock-adapter": "^1.16.0",
-    "babel-core": "^6.26.3",
-    "babel-jest": "^23.4.2",
-    "babel-loader": "^7.1.5",
-    "babel-runtime": "^6.26.0",
+    "babel-jest": "24.8.0",
     "chalk": "^2.4.1",
     "classnames": "^2.2.6",
     "cssnano": "^4.0.5",
     "cypress": "^3.1.5",
-    "enzyme": "^3.4.1",
-    "enzyme-adapter-react-16": "^1.2.0",
+    "enzyme": "^3.10.0",
+    "enzyme-adapter-react-16": "^1.10.0",
     "enzyme-to-json": "^3.3.4",
     "file-loader": "^1.1.11",
     "husky": "^1.0.0-rc.13",
     "isomorphic-ws": "^4.0.1",
-    "jest": "^23.6.0",
+    "jest": "24.7.1",
     "mockserver": "^2.2.0",
     "npm-run-all": "^4.1.5",
     "postcss-autoreset": "^2.0.2",
@@ -103,7 +102,8 @@
     "source-map-explorer": "^1.7.0",
     "ts-jest": "^23.1.3",
     "ts-node": "^7.0.1",
-    "typescript": "^3.0.1",
+    "typescript": "3.4.5",
+    "webpack": "4.29.6",
     "webpack-obfuscator": "^0.18.0"
   },
   "husky": {
@@ -113,14 +113,18 @@
     }
   },
   "prettier": {
-    "trailingComma": "es5",
+    "trailingComma": "es6",
     "tabWidth": 4,
     "printWidth": 130,
     "semi": true,
     "singleQuote": true,
     "tslintIntegration": true,
-    "eslintIntegration": true,
     "stylelintIntegration": true,
     "ignorePath": ".gitignore"
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
