@@ -2,7 +2,7 @@ import { en } from '../../translations/en';
 import { ru } from './ru';
 
 // tslint:disable:no-submodule-imports
-import localeRu = require('react-intl/locale-data/ru');
+import localeRu from 'react-intl/locale-data/ru';
 // tslint:enable
 
 export const customLocaleData = ([...localeRu]);
