@@ -2,21 +2,12 @@ module.exports = {
     'roots': [
         '<rootDir>'
     ],
-    'snapshotSerializers': ['enzyme-to-json/serializer'],
-    'setupTestFrameworkScriptFile': '<rootDir>/enzymeSetup.ts',
     'transform': {
-        '^.+\\.tsx?$': 'ts-jest',
-        '^.+\\.js$': 'babel-jest',
+        '^.+\\.(tsx|ts)$': 'ts-jest',
+        '^.+\\.(jsx|js)$': 'babel-jest',
+        '^.+\\.svg$': '<rootDir>/svgTransform.js',
     },
-    'transformIgnorePatterns': [
-        "<rootDir>/node_modules/(?!lodash-es)"
-    ],
-    'moduleNameMapper': {
-        '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
-            '<rootDir>/__mocks__/fileMock.js',
-        '\\.(css|scss|pcss)$': '<rootDir>/__mocks__/styleMock.js',
-    },
-    'testRegex': '(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$',
+    'testRegex': '(/__tests__/.*|(\\.|/)(test|spec))\\.(tsx|jsx)?$',
     'moduleFileExtensions': [
         'ts',
         'tsx',
@@ -27,6 +18,15 @@ module.exports = {
         'css',
         'scss',
     ],
+    'snapshotSerializers': ['enzyme-to-json/serializer'],
+    'transformIgnorePatterns': [
+        "<rootDir>/node_modules/(?!lodash-es)"
+    ],
+    'moduleNameMapper': {
+        '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
+            '<rootDir>/__mocks__/fileMock.js',
+        '\\.(css|scss|pcss)$': '<rootDir>/__mocks__/styleMock.js',
+    },
     'testURL': "http://localhost",
     'collectCoverage': false,
     'collectCoverageFrom': [
