@@ -3,14 +3,10 @@ import {
 } from '@openware/components';
 import cr from 'classnames';
 import * as React from 'react';
-import {
-    CustomInput,
-} from '../';
-import {
-    EMAIL_REGEX,
-} from '../../helpers';
+import { CustomInput } from '../';
+import { EMAIL_REGEX } from '../../helpers';
 
-interface SignInProps {
+export interface SignInProps {
     labelSignIn?: string;
     labelSignUp?: string;
     emailLabel?: string;
@@ -40,7 +36,7 @@ interface SignInProps {
     changeEmail: (value: string) => void;
 }
 
-class SignInComponent extends React.Component<SignInProps> {
+export class SignInComponent extends React.Component<SignInProps> {
     public render() {
         const {
             email,
@@ -184,8 +180,3 @@ class SignInComponent extends React.Component<SignInProps> {
         }
     };
 }
-
-export {
-    SignInComponent,
-    SignInProps,
-};
