@@ -9,7 +9,7 @@ import {
 } from '../../components';
 import { Beneficiary } from '../../modules';
 
-interface WithdrawProps {
+export interface WithdrawProps {
     currency: string;
     fee: number;
     onClick: (amount: number, total: number, beneficiary: Beneficiary, otpCode: string) => void;
@@ -44,7 +44,7 @@ interface WithdrawState {
     total: number;
 }
 
-class Withdraw extends React.Component<WithdrawProps, WithdrawState> {
+export class Withdraw extends React.Component<WithdrawProps, WithdrawState> {
     public state = {
         amount: '',
         beneficiary: defaultBeneficiary,
@@ -245,9 +245,3 @@ class Withdraw extends React.Component<WithdrawProps, WithdrawState> {
         this.setState({ otpCode });
     };
 }
-
-export {
-    Withdraw,
-    WithdrawProps,
-    WithdrawState,
-};
