@@ -1,9 +1,11 @@
 import Axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
-import WebSocket = require('isomorphic-ws');
 import { Action, Middleware } from 'redux';
 import configureMockStore from 'redux-mock-store';
 import { Config, Cryptobase } from '../api';
+
+// tslint:disable-next-line
+import * as WebSocket from 'ws';
 
 const mockConfig: Config = {
     api: {
